backend/app/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
simplejson
matplotlib
pandas
pydantic-settings
redis==5.0.0